Refactor: move `StudyOwner` to `users` namespace.
