+++ carp.studies.core/src/commonMain/kotlin/dk/cachet/carp/studies/infrastructure/InMemoryStudyRepository.kt
@@ -2,7 +2,7 @@ package dk.cachet.carp.studies.infrastructure
 
 import dk.cachet.carp.common.UUID
 import dk.cachet.carp.studies.domain.Study
-import dk.cachet.carp.studies.domain.StudyOwner
+import dk.cachet.carp.studies.domain.users.StudyOwner
 import dk.cachet.carp.studies.domain.StudyRepository
 import dk.cachet.carp.studies.domain.users.Participant
