+++ carp.studies.core/src/commonTest/kotlin/dk/cachet/carp/studies/infrastructure/StudyOwnerTest.kt
@@ -1,6 +1,6 @@
 package dk.cachet.carp.studies.infrastructure
 
-import dk.cachet.carp.studies.domain.StudyOwner
+import dk.cachet.carp.studies.domain.users.StudyOwner
 import kotlin.test.*
